chore: UserBadge moves into design kit
